Create Adding Digits in numer
parseString(n) does not exist in Java. You should use String.valueOf(n) to convert an integer to a string.

num.length is invalid for a string, you need to use num.length().


(int)num[i] won’t give the digit, it gives the ASCII value of the character. You need to subtract '0'.